Closes-Bug:#1559222 Issue is generated files are shown in git diff command.
As the  *.build.config.js files (* is core/controller/storage/sm) are checkedin, adding to gitignore wont help.
Hence removed from corresponding repos which make files as untracked for which gitignore is applicable.
These file are already listed in gitignore.

In case of controller, storage and server manager build directory consists single file, so removing the single file
from repo deletes the directory, hence we need to create it, while packaging, which is taken care.

Change-Id: Id3866ef4fad222ed5a1f71c1343a114f68ffd99a